Get the calendar for a given month and year

Get the calendar for a given month and year

Usage

get_calendar(
  year = format(Sys.Date(), "%Y"),
  month = format(Sys.Date(), "%m"),
  weekend = FALSE
)

Arguments

year

either an integer or a character of length 1. Must have 4 characters (e.g. '2024' and not '24'). Default is the current year.

month

either an integer or a character of length 1. Must have 1 or 2 characters (e.g. '01' or '1'). Default is the current month.

weekend

a logical. If TRUE keeps Saturdays and Sundays. Default is FALSE.

Value

A data.frame with the following columns:

date: the date of the day (YYYY-MM-DD),

year: the year of the day (integer),

month: the month of the day (integer),

day: the day (integer),

week: the number of week (integer),

weekday: the name of the day of the week (character),

month_name: the name of month (character),

x: position on the x-axis (used by plot_calendar()),

y: position on the y-axis (used by plot_calendar()).

Examples

## Calendar for the current month ----
head(get_calendar())
#>         date year month day week   weekday month_name x y
#> 1 2024-03-01 2024     3   1    9    Friday      March 5 6
#> 2 2024-03-04 2024     3   4   10    Monday      March 1 5
#> 3 2024-03-05 2024     3   5   10   Tuesday      March 2 5
#> 4 2024-03-06 2024     3   6   10 Wednesday      March 3 5
#> 5 2024-03-07 2024     3   7   10  Thursday      March 4 5
#> 6 2024-03-08 2024     3   8   10    Friday      March 5 5

## Calendar for December (current year) ----
head(get_calendar(month = 12))
#>         date year month day week   weekday month_name x y
#> 1 2024-12-02 2024    12   2   49    Monday   December 1 6
#> 2 2024-12-03 2024    12   3   49   Tuesday   December 2 6
#> 3 2024-12-04 2024    12   4   49 Wednesday   December 3 6
#> 4 2024-12-05 2024    12   5   49  Thursday   December 4 6
#> 5 2024-12-06 2024    12   6   49    Friday   December 5 6
#> 6 2024-12-09 2024    12   9   50    Monday   December 1 5

## Calendar for April (current year) ----
head(get_calendar(month = 4))
#>         date year month day week   weekday month_name x y
#> 1 2024-04-01 2024     4   1   14    Monday      April 1 6
#> 2 2024-04-02 2024     4   2   14   Tuesday      April 2 6
#> 3 2024-04-03 2024     4   3   14 Wednesday      April 3 6
#> 4 2024-04-04 2024     4   4   14  Thursday      April 4 6
#> 5 2024-04-05 2024     4   5   14    Friday      April 5 6
#> 6 2024-04-08 2024     4   8   15    Monday      April 1 5

## Calendar for January 1970 ----
head(get_calendar(year = 1970, month = 1))
#>         date year month day week   weekday month_name x y
#> 1 1970-01-01 1970     1   1    0  Thursday    January 4 6
#> 2 1970-01-02 1970     1   2    0    Friday    January 5 6
#> 3 1970-01-05 1970     1   5    1    Monday    January 1 5
#> 4 1970-01-06 1970     1   6    1   Tuesday    January 2 5
#> 5 1970-01-07 1970     1   7    1 Wednesday    January 3 5
#> 6 1970-01-08 1970     1   8    1  Thursday    January 4 5