Bind list of named vectors of variable length into data frame

Based on: https://stackoverflow.com/questions/17308551/

ic_bind_list(x)

Arguments

x

list of named vectors

Examples

ic_bind_list(list(ic_vector(ical_example)))
#>       BEGIN                                     PRODID VERSION  CALSCALE
#> 1 VCALENDAR -//Google Inc//Google Calendar 70.9054//EN     2.0 GREGORIAN
#>    METHOD X-WR-CALNAME X-WR-TIMEZONE          DTSTART            DTEND
#> 1 PUBLISH     atf-test Europe/London 20180809T160000Z 20180809T163000Z
#>            DTSTAMP                                   UID          CREATED
#> 1 20180810T094100Z 1119ejg4vug5758527atjcrqj3@google.com 20180807T133712Z
#>   DESCRIPTION    LAST-MODIFIED LOCATION SEQUENCE    STATUS
#> 1         \\n 20180807T133712Z                 0 CONFIRMED
#>                    SUMMARY TRANSP    END
#> 1 ical programming mission OPAQUE VEVENT
ics_file <- system.file("extdata", "england-and-wales.ics", package = "ical")
ics_raw = readLines(ics_file)
#> Warning: file("") only supports open = "w+" and open = "w+b": using the former
x <- lapply(ic_list(ics_raw), function(x) {
  ic_vector(x)
})
ic_df <- ic_bind_list(x)
head(ic_df)
#> data frame with 0 columns and 0 rows
x <- lapply(ical_outlook, function(x) {
  ic_vector(x)
})
ic_bind_list(x)
#>                                                         DESCRIPTION
#> 1                                           http://2018.erum.io/\\n
#> 2 Ha aaa\\a\\aYau waaa hava praaaaay haard ay aaw ahaa aha aaa aaka
#>                                                                       UID
#> 1 040000008200E00074C5B7101A82E0080000000006245657188FD301000000000000000
#> 2 040000008200E00074C5B7101A82E00800000000F0CCE6F58CC2D301000000000000000
#>    01000000096B0CD2B1133C746B52B5B0B1F12B731
#> 1  01000000096B0CD2B1133C746B52B5B0B1F12B731
#> 2                                       <NA>
#>                                                      SUMMARY DTSTART;VALUE=DATE
#> 1 In Budapest for European R Users Meeting (eRum) conference           20180514
#> 2              Caacaaad: aaa aaka shara - Laads aauach avaaa               <NA>
#>   DTEND;VALUE=DATE  CLASS PRIORITY          DTSTAMP      TRANSP    STATUS
#> 1         20180518 PUBLIC        5 20180814T134259Z TRANSPARENT CONFIRMED
#> 2             <NA> PUBLIC        5 20180814T134259Z TRANSPARENT CONFIRMED
#>   SEQUENCE                                           LOCATION
#> 1        0                                                   
#> 2        6 Leeds City Centre - time and venue to be confirmed
#>   X-MICROSOFT-CDO-APPT-SEQUENCE X-MICROSOFT-CDO-BUSYSTATUS
#> 1                             0                       FREE
#> 2                             6                       FREE
#>   X-MICROSOFT-CDO-INTENDEDSTATUS X-MICROSOFT-CDO-ALLDAYEVENT
#> 1                           BUSY                        TRUE
#> 2                           BUSY                       FALSE
#>   X-MICROSOFT-CDO-IMPORTANCE X-MICROSOFT-CDO-INSTTYPE
#> 1                          1                        0
#> 2                          1                        0
#>   X-MICROSOFT-DONOTFORWARDMEETING X-MICROSOFT-DISALLOW-COUNTER
#> 1                           FALSE                        FALSE
#> 2                           FALSE                        FALSE
#>    010000000FB969A27B8680C419BAD7A7656340914 DTSTART;TZID=GMT Standard Time
#> 1                                       <NA>                           <NA>
#> 2  010000000FB969A27B8680C419BAD7A7656340914                20180515T100000
#>   DTEND;TZID=GMT Standard Time
#> 1                         <NA>
#> 2              20180515T160000