Create ical object from properties_core inputs

ic_event(
  uid = ic_guid(),
  start_time = as.POSIXct(round.POSIXt(Sys.time(), units = "hours")),
  end_time = 1,
  format = "%Y-%m-%d %H:%M",
  summary = "ical event",
  more_properties = FALSE,
  event_properties = calendar::properties
)

Arguments

uid
the unique id of the event, by default generated by ic_uid()

start_time
start time, by default the start time plus one hour

end_time
a number representing the number of hours after start_time or an datetime object (of class POSIXct) when the event ends. By default set to 1, meaning 1 hour after start_time.

format
required if start_time and end_time are vectors and are not of datetime format "\

summaryshort outline of the event

more_propertiesadd placeholder columns for properties in addition to properties_core, dy default FALSE

event_propertiesnamed vector of additional properties to include. By default These include names stored in the data object properties (a packaged dataset), containing NAs to be subsequently populated.

Format

in case of start_time and end_time being character values, a format must be provided.